feat: rename search utils
