+++ packages/ui/src/hooks/useListOption.js
@@ -1,5 +1,5 @@
 import { useHistory } from "react-router-dom";
-import { addSearchToLocalStorage } from "../utils/searchUtil";
+import { addSearchToLocalStorage } from "../utils/searchUtils";
 
 function useListOption() {
   let history = useHistory();
