+++ packages/ui/src/utils/searchUtils.js
